Show OMP calls in physical time.

+++ src/commevent.h
@@ -137,7 +137,7 @@ public:
     virtual void initialize_strides(QList<CommEvent *> * stride_events,
                                     QList<CommEvent *> * recv_events)=0;
     virtual void update_strides() { return; }
-    virtual void set_reorder_strides(QMap<int, QList<CommEvent *> *> * stride_map,
+    virtual void set_reorder_strides(QMap<unsigned long, QList<CommEvent *> *> * stride_map,
                                      int offset, CommEvent * last = NULL, int debug = -1)
                                     { Q_UNUSED(stride_map); Q_UNUSED(offset); Q_UNUSED(last); Q_UNUSED(debug); return; }
     virtual void initialize_basic_strides(QSet<CollectiveRecord *> * collectives)=0;
