+++ src/event.h
@@ -15,7 +15,7 @@ class Event
 {
 public:
     Event(unsigned long long _enter, unsigned long long _exit, int _function,
-          int _entity, int _pe);
+          unsigned long _entity, int _pe);
     ~Event();
 
     // Based on enter time
@@ -51,7 +51,7 @@ public:
     unsigned long long enter;
     unsigned long long exit;
     int function;
-    int entity;
+    unsigned long entity;
     int pe;
     int depth;
