+++ src/overviewvis.cpp
@@ -226,7 +226,7 @@ void OverviewVis::processVis()
     for (QList<Partition *>::Iterator part = trace->partitions->begin();
          part != trace->partitions->end(); ++part)
     {
-        for (QMap<int, QList<CommEvent *> *>::Iterator event_list
+        for (QMap<unsigned long, QList<CommEvent *> *>::Iterator event_list
              = (*part)->events->begin();
              event_list != (*part)->events->end(); ++event_list)
         {
